Big refactoring of transport related code. o Move to a Channel API. See the Channel interface o Create MessageExchange classes which control the message exchange patterns. o Commit some basic WS-Addressing stuff which will be used for a new MEP
NOTE: the new client stuff isn't exactly done yet, but rather than wait, I wanted to get this in CVS so people can start playing with it. In the meantime I left the old stuff in.  Hopefully you'll see how the new Channel and MessageExchange stuff is leading to a much nicer client api.


git-svn-id: http://svn.codehaus.org/xfire/trunk/xfire@401 c7d0bf07-ec0d-0410-b2cc-d48fa9be22ba