A significant clean up of the Transport/Channel model and the Handler chains.
o WS-Addressing is now supported
o Channels can no longer be associated with services. They expect to be
able to determine the service by the URI or WS-Addressing
o Handlers chains are now dynamic in nature. Additional handlers can be
added at any time.
o Handlers are grouped by "Phases" - this is an Axis 2.0 feature that I liked
and decided to integrate.


git-svn-id: http://svn.codehaus.org/xfire/trunk/xfire@556 c7d0bf07-ec0d-0410-b2cc-d48fa9be22ba